requirements.txt
# Core LangChain and Gemini
langgraph
langchain
langchain_google_genai

# Data handling and visualization
pandas
matplotlib

# Environment variable management
python-dotenv
